Extract Simple functionality out into SimpleEntity
